move checkbox-node
