+++ packages/vlossom/src/components/index.ts
@@ -1,44 +1,43 @@
-export { type VsAccordionStyleSet } from './vs-accordion/types';
+export type { VsAccordionStyleSet } from './vs-accordion/types';
 export { default as VsAccordion } from './vs-accordion/VsAccordion.vue';
 
-export { type VsAvatarStyleSet } from './vs-avatar/types';
+export type { VsAvatarStyleSet } from './vs-avatar/types';
 export { default as VsAvatar } from './vs-avatar/VsAvatar.vue';
 
-export { type VsBlockStyleSet } from './vs-block/types';
+export type { VsBlockStyleSet } from './vs-block/types';
 export { default as VsBlock } from './vs-block/VsBlock.vue';
 
-export { type VsButtonStyleSet } from './vs-button/types';
+export type { VsButtonStyleSet } from './vs-button/types';
 export { default as VsButton } from './vs-button/VsButton.vue';
 
-export { type VsCheckboxStyleSet } from './vs-checkbox/types';
+export type { VsCheckboxNodeStyleSet, VsCheckboxStyleSet, VsCheckboxSetStyleSet } from './vs-checkbox/types';
+export { default as VsCheckboxNode } from './vs-checkbox/VsCheckboxNode.vue';
 export { default as VsCheckbox } from './vs-checkbox/VsCheckbox.vue';
+export { default as VsCheckboxSet } from './vs-checkbox/VsCheckboxSet.vue';
 
-export { type VsCheckboxSetStyleSet } from './vs-checkbox-set/types';
-export { default as VsCheckboxSet } from './vs-checkbox-set/VsCheckboxSet.vue';
-
-export { type VsChipStyleSet } from './vs-chip/types';
+export type { VsChipStyleSet } from './vs-chip/types';
 export { default as VsChip } from './vs-chip/VsChip.vue';
 
 export { default as VsContainer } from './vs-container/VsContainer.vue';
 
-export { type VsDividerStyleSet } from './vs-divider/types';
+export type { VsDividerStyleSet } from './vs-divider/types';
 export { default as VsDivider } from './vs-divider/VsDivider.vue';
 
-export { type VsDrawerStyleSet } from './vs-drawer/types';
+export type { VsDrawerStyleSet } from './vs-drawer/types';
 export { default as VsDrawer } from './vs-drawer/VsDrawer.vue';
 
-export { type VsFileInputStyleSet } from './vs-file-input/types';
+export type { VsFileInputStyleSet } from './vs-file-input/types';
 export { default as VsFileInput } from './vs-file-input/VsFileInput.vue';
 
-export { type VsFooterStyleSet } from './vs-footer/types';
+export type { VsFooterStyleSet } from './vs-footer/types';
 export { default as VsFooter } from './vs-footer/VsFooter.vue';
 
 export { default as VsForm } from './vs-form/VsForm.vue';
 
-export { type VsHeaderStyleSet } from './vs-header/types';
+export type { VsHeaderStyleSet } from './vs-header/types';
 export { default as VsHeader } from './vs-header/VsHeader.vue';
 
-export { type VsImageStyleSet } from './vs-image/types';
+export type { VsImageStyleSet } from './vs-image/types';
 export { default as VsImage } from './vs-image/VsImage.vue';
 
 export { default as VsIndexItem } from './vs-index-view/VsIndexItem.vue';
@@ -49,70 +48,70 @@ export { default as VsInput } from './vs-input/VsInput.vue';
 
 export { default as VsInputWrapper } from './vs-input-wrapper/VsInputWrapper.vue';
 
-export { type VsLabelValueStyleSet } from './vs-label-value/types';
+export type { VsLabelValueStyleSet } from './vs-label-value/types';
 export { default as VsLabelValue } from './vs-label-value/VsLabelValue.vue';
 
 export { default as VsLayout } from './vs-layout/VsLayout.vue';
 
-export { type VsLoadingStyleSet } from './vs-loading/types';
+export type { VsLoadingStyleSet } from './vs-loading/types';
 export { default as VsLoading } from './vs-loading/VsLoading.vue';
 
 export { default as VsMessage } from './vs-message/VsMessage.vue';
 
 export { default as VsModal } from './vs-modal/VsModal.vue';
 
-export { type VsNoticeStyleSet } from './vs-notice/types';
+export type { VsNoticeStyleSet } from './vs-notice/types';
 export { default as VsNotice } from './vs-notice/VsNotice.vue';
 
-export { type VsPageStyleSet } from './vs-page/types';
+export type { VsPageStyleSet } from './vs-page/types';
 export { default as VsPage } from './vs-page/VsPage.vue';
 
-export { type VsPaginationStyleSet } from './vs-pagination/types';
+export type { VsPaginationStyleSet } from './vs-pagination/types';
 export { default as VsPagination } from './vs-pagination/VsPagination.vue';
 
-export { type VsProgressStyleSet } from './vs-progress/types';
+export type { VsProgressStyleSet } from './vs-progress/types';
 export { default as VsProgress } from './vs-progress/VsProgress.vue';
 
-export { type VsRadioStyleSet } from './vs-radio/types';
+export type { VsRadioStyleSet } from './vs-radio/types';
 export { default as VsRadio } from './vs-radio/VsRadio.vue';
 
-export { type VsRadioSetStyleSet } from './vs-radio-set/types';
+export type { VsRadioSetStyleSet } from './vs-radio-set/types';
 export { default as VsRadioSet } from './vs-radio-set/VsRadioSet.vue';
 
 export { default as VsResponsive } from './vs-responsive/VsResponsive.vue';
 
-export { type VsSectionStyleSet } from './vs-section/types';
+export type { VsSectionStyleSet } from './vs-section/types';
 export { default as VsSection } from './vs-section/VsSection.vue';
 
-export { type VsSelectStyleSet } from './vs-select/types';
+export type { VsSelectStyleSet } from './vs-select/types';
 export { default as VsSelect } from './vs-select/VsSelect.vue';
 
-export { type VsSkeletonStyleSet } from './vs-skeleton/types';
+export type { VsSkeletonStyleSet } from './vs-skeleton/types';
 export { default as VsSkeleton } from './vs-skeleton/VsSkeleton.vue';
 
-export { type VsStepperStyleSet } from './vs-stepper/types';
+export type { VsStepperStyleSet } from './vs-stepper/types';
 export { default as VsStepper } from './vs-stepper/VsStepper.vue';
 
-export { type VsSwitchStyleSet } from './vs-switch/types';
+export type { VsSwitchStyleSet } from './vs-switch/types';
 export { default as VsSwitch } from './vs-switch/VsSwitch.vue';
 
 export { SortType } from './vs-table/types';
 export type { VsTableStyleSet, TableHeader, TableItem, TableParams, TableRow, TableFilter } from './vs-table/types';
 export { default as VsTable } from './vs-table/VsTable.vue';
 
-export { type VsTabsStyleSet } from './vs-tabs/types';
+export type { VsTabsStyleSet } from './vs-tabs/types';
 export { default as VsTabs } from './vs-tabs/VsTabs.vue';
 
-export { type VsTextWrapStyleSet } from './vs-text-wrap/types';
+export type { VsTextWrapStyleSet } from './vs-text-wrap/types';
 export { default as VsTextWrap } from './vs-text-wrap/VsTextWrap.vue';
 
-export { type VsTextareaStyleSet } from './vs-textarea/types';
+export type { VsTextareaStyleSet } from './vs-textarea/types';
 export { default as VsTextarea } from './vs-textarea/VsTextarea.vue';
 
-export { type VsThemeButtonStyleSet } from './vs-theme-button/types';
+export type { VsThemeButtonStyleSet } from './vs-theme-button/types';
 export { default as VsThemeButton } from './vs-theme-button/VsThemeButton.vue';
 
-export { type VsToastStyleSet } from './vs-toast/types';
+export type { VsToastStyleSet } from './vs-toast/types';
 
-export { type VsTooltipStyleSet } from './vs-tooltip/types';
+export type { VsTooltipStyleSet } from './vs-tooltip/types';
 export { default as VsTooltip } from './vs-tooltip/VsTooltip.vue';
