--- packages/vlossom/src/components/vs-checkbox-set/types.ts
@@ -1,6 +1,0 @@
-import type { VsCheckboxNodeStyleSet } from '@/nodes';
-
-export interface VsCheckboxSetStyleSet extends VsCheckboxNodeStyleSet {
-    checkboxGap?: string;
-    flexWrap?: string;
-}
