+++ packages/vlossom/src/components/vs-checkbox/VsCheckboxNode.vue
