+++ packages/vlossom/src/components/vs-checkbox/VsCheckboxSet.vue
@@ -62,7 +62,8 @@ import { getInputProps, getInputOptionProps, getResponsiveProps } from '@/models
 import { VsComponent, VsNode, type ColorScheme } from '@/declaration';
 import { utils } from '@/utils';
 import VsInputWrapper from '@/components/vs-input-wrapper/VsInputWrapper.vue';
-import { VsCheckboxNode, VsCheckboxNodeStyleSet } from '@/nodes';
+import VsCheckboxNode from './VsCheckboxNode.vue';
+import { VsCheckboxNodeStyleSet } from './types';
 import { useVsCheckboxSetRules } from './vs-checkbox-set-rules';
 
 import type { VsCheckboxSetStyleSet } from './types';
