+++ packages/vlossom/src/components/vs-checkbox/types.ts
@@ -1,3 +1,16 @@
-import type { VsCheckboxNodeStyleSet } from '@/nodes';
+export interface VsCheckboxNodeStyleSet {
+    borderRadius?: string;
+    checkboxColor?: string;
+    checkboxSize?: string;
+    height?: string;
+    labelFontColor?: string;
+    labelFontSize?: string;
+    selectedLabelFontColor?: string;
+}
 
 export interface VsCheckboxStyleSet extends VsCheckboxNodeStyleSet {}
+
+export interface VsCheckboxSetStyleSet extends VsCheckboxNodeStyleSet {
+    checkboxGap?: string;
+    flexWrap?: string;
+}
