--- packages/vlossom/src/nodes/vs-checkbox-node/types.ts
@@ -1,9 +1,0 @@
-export interface VsCheckboxNodeStyleSet {
-    borderRadius?: string;
-    checkboxColor?: string;
-    checkboxSize?: string;
-    height?: string;
-    labelFontColor?: string;
-    labelFontSize?: string;
-    selectedLabelFontColor?: string;
-}
